requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
pandas
pycountry
plotly
